+++ src/chain_sync/tipset_syncer.rs
@@ -1565,8 +1565,8 @@ async fn check_block_messages<DB: Blockstore + Send + Sync + 'static>(
             .map_err(|e| TipsetRangeSyncerError::ResolvingAddressFromMessage(e.to_string()))?;
         // SecP256K1 Signature validation
         msg.signature
-            .verify(&msg.message().cid().to_bytes(), &key_addr)
-            .map_err(TipsetRangeSyncerError::MessageSignatureInvalid)?;
+            .authenticate_msg(eth_chain_id, msg, &key_addr)
+            .map_err(|e| TipsetRangeSyncerError::MessageSignatureInvalid(e.to_string()))?;
     }
 
     // Validate message root from header matches message root