+++ src/eth/eip_1559_transaction.rs
@@ -139,9 +139,25 @@ impl EthEip1559TxArgs {
         from: Address,
         eth_chain_id: EthChainId,
     ) -> anyhow::Result<SignedMessage> {
-        ensure!(self.chain_id != eth_chain_id, "Invalid chain id");
+        let message = self.get_unsigned_message(from, eth_chain_id)?;
+        let signature = self.signature()?;
+        Ok(SignedMessage { message, signature })
+    }
+
+    /// Constructs an unsigned message using EIP-1559 transaction args
+    pub fn get_unsigned_message(
+        &self,
+        from: Address,
+        eth_chain_id: EthChainId,
+    ) -> anyhow::Result<Message> {
+        ensure!(
+            self.chain_id == eth_chain_id,
+            "Invalid chain id, expected {}, got {}",
+            self.chain_id,
+            eth_chain_id
+        );
         let method_info = get_filecoin_method_info(&self.to, &self.input)?;
-        let message = Message {
+        Ok(Message {
             version: 0,
             from,
             to: method_info.to,
@@ -152,9 +168,7 @@ impl EthEip1559TxArgs {
             gas_limit: self.gas_limit,
             gas_fee_cap: self.max_fee_per_gas.clone().into(),
             gas_premium: self.max_priority_fee_per_gas.clone().into(),
-        };
-        let signature = self.signature()?;
-        Ok(SignedMessage { message, signature })
+        })
     }
 }